+++ src/main/java/com/fullmugu/nanumeal/api/dto/user/InputUserInfoRequestDto.java
@@ -1,4 +1,4 @@
-package com.fullmugu.nanumeal.api.dto;
+package com.fullmugu.nanumeal.api.dto.user;
 
 import lombok.Data;
