+++ src/main/java/com/fullmugu/nanumeal/api/service/user/UserService.java
@@ -1,7 +1,7 @@
 package com.fullmugu.nanumeal.api.service.user;
 
 
-import com.fullmugu.nanumeal.api.dto.InputUserInfoRequestDto;
+import com.fullmugu.nanumeal.api.dto.user.InputUserInfoRequestDto;
 import com.fullmugu.nanumeal.api.dto.user.UserDTO;
 import com.fullmugu.nanumeal.api.entity.user.User;
